Changes have been tested. This build is ready to push to develop, although significant refactoring is recommended before pushing to main

+++ Business/ProfileLogic.cs
@@ -1,18 +1,7 @@
-﻿using OpenAICustomFunctionCallingAPI.Client;
-using OpenAICustomFunctionCallingAPI.Controllers.DTOs;
-using Newtonsoft.Json.Linq;
-using OpenAICustomFunctionCallingAPI.Host.Config;
+﻿using OpenAICustomFunctionCallingAPI.Controllers.DTOs;
 //using OpenAICustomFunctionCallingAPI.DAL;
-using Azure;
-using Microsoft.Data.SqlClient;
 using OpenAICustomFunctionCallingAPI.DAL;
-using System.Web.Mvc;
-using System.Net;
-using System.Net.Mail;
-using Microsoft.AspNetCore.Mvc;
-using Nest;
 using OpenAICustomFunctionCallingAPI.DAL.DTOs;
-using System.Reflection.Metadata;
 using OpenAICustomFunctionCallingAPI.API.DTOs.ClientDTOs.ToolDTOs;
 
 namespace OpenAICustomFunctionCallingAPI.Business.ProfileLogic
@@ -42,35 +31,27 @@ namespace OpenAICustomFunctionCallingAPI.Business.ProfileLogic
         public async Task<APIProfileDTO> GetProfile(string name)
         {
             var dbProfile = await _profileDb.GetByNameWithToolsAsync(name);
-            if (dbProfile != null)
+            if (dbProfile is not null)
             {
                 // package this into a seperate method (same one as in GetAllProfiles())
                 var profileToolDTOs = await _profileToolsDb.GetToolAssociationsAsync(dbProfile.Id);
                 dbProfile.Tools = new List<ToolDTO>();
-                foreach (var association in profileToolDTOs)
-                {
-                    var tool = await _toolDb.GetToolByIdAsync(association.ToolID);
-                    dbProfile.Tools.Add(tool);
-                }
+                foreach (var association in profileToolDTOs) dbProfile.Tools.Add(await _toolDb.GetToolByIdAsync(association.ToolID));
                 return dbProfile;
             }
             else
             {
                 var apiProfile = await _profileDb.GetByNameAsync(name);
-                if (apiProfile != null)
-                {
-                    return new APIProfileDTO(apiProfile);
-                }
+                if (apiProfile != null) return new APIProfileDTO(apiProfile);
             }
             return null;
         }
 
-        // might need to return as pageable if db starts getting big
         public async Task<IEnumerable<APIProfileDTO>> GetAllProfiles()
         {
             var response = await _profileDb.GetAllAsync();
             var apiResponseList = new List<APIProfileDTO>();
-            if (response != null)
+            if (response is not null)
             {
                 foreach (var profile in response)
                 {
@@ -89,93 +70,59 @@ namespace OpenAICustomFunctionCallingAPI.Business.ProfileLogic
             return apiResponseList;
         }
 
+        // refactor this
         public async Task<string> CreateOrUpdateProfile(APIProfileDTO profileDto)
         {
             var errorMessage = _validationLogic.ValidateAPIProfile(profileDto); // move to controller
-            if (errorMessage != null)
-            {
-                return errorMessage;
-            }
-
+            if (errorMessage is not null) return errorMessage;
             var existingProfile = await _profileDb.GetByNameAsync(profileDto.Name);
             var success = true;
-            if (existingProfile != null)
+            if (existingProfile is not null)
             {
                 var updateProfileDto = new DbProfileDTO(existingProfile, profileDto);
                 var rows = await _profileDb.UpdateAsync(existingProfile, updateProfileDto);
-                if (rows != 1)
-                {
-                    success = false;
-                }
+                if (rows != 1) success = false;
             }
             else
             {
                 var updateProfileDto = new DbProfileDTO(null, profileDto);
                 var newTool = await _profileDb.AddAsync(updateProfileDto);
-                if (newTool == null)
-                {
-                    success = false;
-                }
+                if (newTool is null) success = false;
             }
 
-            if (profileDto.Tools != null && profileDto.Tools.Count > 0)
+            if (profileDto.Tools is not null && profileDto.Tools.Count > 0)
             {
-                if (existingProfile == null)
-                {
-                    await AddOrUpdateProfileTools(profileDto, null);
-                }
-                else
-                {
-                    await AddOrUpdateProfileTools(profileDto, new APIProfileDTO(existingProfile));
-                }
+                if (existingProfile is null) await AddOrUpdateProfileTools(profileDto, null);
+                else await AddOrUpdateProfileTools(profileDto, new APIProfileDTO(existingProfile));
             }
-
             var existingProfileWithTools = await _profileDb.GetByNameWithToolsAsync(profileDto.Name);
-            if (success && existingProfileWithTools != null)
-            {
-                success = await AddOrUpdateProfileTools(profileDto, existingProfileWithTools);
-                
-            }
-
-            if (!success)
-            {
-                return $"Something went wrong...";
-            }
+            if (success && existingProfileWithTools is not null) success = await AddOrUpdateProfileTools(profileDto, existingProfileWithTools);
+            if (!success) return $"Something went wrong...";
             return null;
         }
 
-        // could use some refactoring
+        // could also use some refactoring
         public async Task<string> DeleteProfile(string name)
         {
             var profileDto = await _profileDb.GetByNameWithToolsAsync(name);
             int rows;
-            if (profileDto != null)
+            if (profileDto is not null)
             {
-                if (profileDto.Tools != null && profileDto.Tools.Count > 0)
-                {
-                    await _profileToolsDb.DeleteAllProfileAssociationsAsync(profileDto.Id);
-                }
+                if (profileDto.Tools is not null && profileDto.Tools.Count > 0) await _profileToolsDb.DeleteAllProfileAssociationsAsync(profileDto.Id);
                 var dbProfileDTO = new DbProfileDTO()
                 {
                     Id = profileDto.Id,
                     Name = profileDto.Name
                 };
-                rows = await _profileDb.DeleteAsync(dbProfileDTO); 
-                
+                rows = await _profileDb.DeleteAsync(dbProfileDTO);
             }
             else
             {
                 var toolLessProfile = await _profileDb.GetByNameAsync(name);
-                if (toolLessProfile == null)
-                {
-                    return $"No profile with the name '{name}' was found.";
-                }
+                if (toolLessProfile is null) return $"No profile with the name '{name}' was found.";
                 rows = await _profileDb.DeleteAsync(toolLessProfile);
             }
-            if (rows != 1)
-            {
-                return "something went wrong while deleting the tool. Please note that the profile associations were deleted successfully.";
-            }
+            if (rows != 1) return "something went wrong while deleting the tool. Please note that the profile associations were deleted successfully.";
             return null;
         }
 
@@ -184,7 +131,7 @@ namespace OpenAICustomFunctionCallingAPI.Business.ProfileLogic
         {
             var toolIds = new List<int>();
             await _profileToolsDb.DeleteAllProfileAssociationsAsync(existingProfile.Id);
-            if (profileDto.Tools != null && profileDto.Tools.Count > 0)
+            if (profileDto.Tools is not null && profileDto.Tools.Count > 0)
             {
                 await CreateOrUpdateTools(profileDto.Tools);
                 foreach (var tool in profileDto.Tools)
@@ -192,7 +139,6 @@ namespace OpenAICustomFunctionCallingAPI.Business.ProfileLogic
                     var dbTool = await _toolDb.GetToolByNameAsync(tool.Function.Name);
                     toolIds.Add(dbTool.Id);
                 }
-
                 await _profileToolsDb.AddAssociationsByProfileIdAsync(existingProfile.Id, toolIds);
             }
             return true;
@@ -215,14 +161,18 @@ namespace OpenAICustomFunctionCallingAPI.Business.ProfileLogic
             return returnList;
         }
 
+        public async Task<List<string>> GetProfileToolAssociations(string name)
+        {
+            var profileNames = await _toolDb.GetProfileToolsAsync(name);
+            if (profileNames.Count > 0) return profileNames;
+            else return null;
+        }
+
         public async Task<List<string>> GetToolProfileAssociations(string name)
         {
             var profileNames = await _toolDb.GetToolProfilesAsync(name);
-            if (profileNames.Count > 0)
-            {
-                return profileNames;
-            }
-            return null;
+            if (profileNames.Count > 0) return profileNames;
+            else return null;
         }
 
         public async Task<string> CreateOrUpdateTools(List<ToolDTO> toolList)
@@ -231,10 +181,7 @@ namespace OpenAICustomFunctionCallingAPI.Business.ProfileLogic
             foreach (var tool in toolList)
             {
                 var errorMessage = _validationLogic.ValidateTool(tool);
-                if (errorMessage != null)
-                {
-                    return errorMessage;
-                }
+                if (errorMessage is not null) return errorMessage;
             }
             foreach (var tool in toolList)
             {
@@ -244,45 +191,66 @@ namespace OpenAICustomFunctionCallingAPI.Business.ProfileLogic
                 {
                     var existingTool = new DbToolDTO(existingToolDTO);
                     await _toolDb.UpdateAsync(existingTool, dbToolDTO);
-                    await AddOrUpdateToolProperties(existingToolDTO, tool.Function.Parameters.Properties);
+                    await AddOrUpdateToolProperties(existingToolDTO, tool.Function.Parameters.properties);
                 }
                 else
                 {
                     var newTool = await _toolDb.AddAsync(dbToolDTO);
                     var newToolDTO = new ToolDTO(newTool, null);
-                    await AddOrUpdateToolProperties(newToolDTO, tool.Function.Parameters.Properties);
+                    await AddOrUpdateToolProperties(newToolDTO, tool.Function.Parameters.properties);
                 }
             }
             return null;
         }
 
-        public async Task<string> AddToolAssociations(string name, List<string> profiles)
+        public async Task<string> AddToolToProfiles(string name, List<string> profiles)
         {
             var tool = await _toolDb.GetByNameAsync(name);
-            if (tool != null)
+            if (tool is not null)
             {
                 var success = await _profileToolsDb.AddAssociationsByToolIdAsync(tool.Id, profiles);
-                if (success)
-                {
-                    return null;
-                }
-                else
-                {
-                    return "something went wrong...";
-                }
+                if (success) return null;
+                else return "something went wrong...";
             }
             return $"No tool with the name {name} exists.";
+        }
+
+        public async Task<string> AddProfileToTools(string name, List<string> tools)
+        {
+            var toolIDs = new List<int>();
+            foreach (var toolName in tools)
+            {
+                var tool = await _toolDb.GetByNameAsync(toolName);
+                if (tool is null) return $"No tool '{toolName}' found.";
+                toolIDs.Add(tool.Id);
+            }
+            var profile = await _profileDb.GetByNameAsync(name);
+            if (profile is not null && toolIDs.Count > 0)
+            {
+                var success = await _profileToolsDb.AddAssociationsByProfileIdAsync(profile.Id, toolIDs);
+                if (success) return null;
+                else return $"No profile {profile.Name} found.";
+            }
+            return $"The tools or profile in the request body were not found.";
         }
 
         public async Task<string> DeleteToolAssociations(string name, List<string> profiles)
         {
             var tool = await _toolDb.GetByNameAsync(name);
-            if (tool != null)
+            if (tool is not null)
             {
-                foreach (var profile in profiles)
-                {
-                    await _profileToolsDb.DeleteAssociationAsync(tool.Id, profile);
-                }
+                foreach (var profile in profiles) await _profileToolsDb.DeleteToolAssociationAsync(tool.Id, profile);
+                return null;
+            }
+            return $"No tool with the name {name} exists.";
+        }
+
+        public async Task<string> DeleteProfileAssociations(string name, List<string> tools)
+        {
+            var profile = await _profileDb.GetByNameAsync(name);
+            if (profile is not null)
+            {
+                foreach (var tool in tools) await _profileToolsDb.DeleteProfileAssociationAsync(profile.Id, tool);
                 return null;
             }
             return $"No tool with the name {name} exists.";
@@ -291,9 +259,9 @@ namespace OpenAICustomFunctionCallingAPI.Business.ProfileLogic
         public async Task<bool> DeleteTool(string name)
         {
             var existingTool = await _toolDb.GetToolByNameAsync(name);
-            if (existingTool != null)
+            if (existingTool is not null)
             {
-                foreach (var property in existingTool.Function.Parameters.Properties)
+                foreach (var property in existingTool.Function.Parameters.properties)
                 {
                     var propertyDTO = new DbPropertyDTO(property.Key, property.Value);
                     await _propertyDb.DeleteAsync(propertyDTO);
@@ -307,14 +275,10 @@ namespace OpenAICustomFunctionCallingAPI.Business.ProfileLogic
         public async Task<bool> AddOrUpdateToolProperties(ToolDTO existingTool, Dictionary<string, PropertyDTO> newProperties)
         {
             var existingProperties = await _propertyDb.GetToolProperties(existingTool.Id);
-            foreach (var property in existingProperties)
-            {
-                await _propertyDb.DeleteAsync(property);
-            }
-
+            foreach (var property in existingProperties) await _propertyDb.DeleteAsync(property);
             foreach (var property in newProperties)
             {
-                property.Value.Id = existingTool.Id;
+                property.Value.id = existingTool.Id;
                 await _propertyDb.AddAsync(new DbPropertyDTO(property.Key, property.Value));
             }
             return true;
