+++ Controllers/ToolController.cs
@@ -35,16 +35,10 @@ namespace OpenAICustomFunctionCallingAPI.Controllers
         {
             try
             {
-                if (string.IsNullOrEmpty(name))
-                {
-                    return BadRequest($"Invalid request.Please check the route parameter for the profile name: {name}.");
-                }
+                if (string.IsNullOrEmpty(name)) return BadRequest($"Invalid request.Please check the route parameter for the profile name: {name}.");
                 var tool = await _profileLogic.GetTool(name);
-                if (tool == null)
-                {
-                    return NotFound($"No tool with the name {name} exists");
-                }
-                return Ok(tool);
+                if (tool == null) return NotFound($"No tool with the name {name} exists");
+                else return Ok(tool);
             }
             catch (HttpRequestException ex)
             {
@@ -65,11 +59,8 @@ namespace OpenAICustomFunctionCallingAPI.Controllers
             try
             {
                 var tools = await _profileLogic.GetAllTools();
-                if (tools == null || tools.Count() < 1)
-                {
-                    return NotFound($"No tools exist. Make a post request to add some.");
-                }
-                return Ok(tools);
+                if (tools == null || tools.Count() < 1) return NotFound($"No tools exist. Make a post request to add some.");
+                else return Ok(tools);
             }
             catch (HttpRequestException ex)
             {
@@ -88,16 +79,10 @@ namespace OpenAICustomFunctionCallingAPI.Controllers
         {
             try
             {
-                if (string.IsNullOrEmpty(name))
-                {
-                    return BadRequest($"Invalid request.Please check the route parameter for the profile name: {name}.");
-                }
+                if (string.IsNullOrEmpty(name)) return BadRequest($"Invalid request.Please check the route parameter for the profile name: {name}.");
                 var tool = await _profileLogic.GetToolProfileAssociations(name);
-                if (tool == null)
-                {
-                    return NotFound($"The tool '{name}' is not associated with any profiles, or does not exist.");
-                }
-                return Ok(tool);
+                if (tool == null) return NotFound($"The tool '{name}' is not associated with any profiles, or does not exist.");
+                else return Ok(tool);
             }
             catch (HttpRequestException ex)
             {
@@ -117,11 +102,8 @@ namespace OpenAICustomFunctionCallingAPI.Controllers
             try
             {
                 var errorMessage = await _profileLogic.CreateOrUpdateTools(toolList);
-                if (errorMessage != null)
-                {
-                    return BadRequest(errorMessage);
-                }
-                return NoContent();
+                if (errorMessage != null) return BadRequest(errorMessage);
+                else return NoContent();
             }
             catch (HttpRequestException ex)
             {
@@ -129,7 +111,6 @@ namespace OpenAICustomFunctionCallingAPI.Controllers
             }
             catch (Exception)
             {
-                // Logging and transient exception handling here
                 throw;
             }
         }
@@ -138,25 +119,13 @@ namespace OpenAICustomFunctionCallingAPI.Controllers
         [Route("associate/{name}")]
         public async Task<IActionResult> AddToolToProfiles([FromRoute] string name, List<string> profiles)
         {
-            // prevent profiles that don't exist from being added?
-
             try
             {
-                if (string.IsNullOrEmpty(name))
-                {
-                    return BadRequest($"Invalid request.Please check the route parameter for the profile name: {name}.");
-                }
-                if (profiles == null || profiles.Count < 1)
-                {
-                    return BadRequest($"Invalid request.'Profiles' property cannot be null or empty: {profiles}.");
-                }
-
-                var errorMessage = await _profileLogic.AddToolAssociations(name, profiles);
-                if (errorMessage == null)
-                {
-                    return Ok(await _profileLogic.GetToolProfileAssociations(name));
-                }
-                return NotFound(errorMessage);
+                if (string.IsNullOrEmpty(name)) return BadRequest($"Invalid request.Please check the route parameter for the profile name: {name}.");
+                if (profiles == null || profiles.Count < 1) return BadRequest($"Invalid request.'Profiles' property cannot be null or empty: {profiles}.");
+                var errorMessage = await _profileLogic.AddToolToProfiles(name, profiles);
+                if (errorMessage == null) return Ok(await _profileLogic.GetToolProfileAssociations(name));
+                else return NotFound(errorMessage);
             }
             catch (HttpRequestException ex)
             {
@@ -175,21 +144,11 @@ namespace OpenAICustomFunctionCallingAPI.Controllers
         {
             try
             {
-                if (string.IsNullOrEmpty(name))
-                {
-                    return BadRequest($"Invalid request.Please check the route parameter for the profile name: {name}.");
-                }
-                if (profiles == null || profiles.Count < 1)
-                {
-                    return BadRequest($"Invalid request.'Profiles' property cannot be null or empty: {profiles}.");
-                }
-
+                if (string.IsNullOrEmpty(name)) return BadRequest($"Invalid request.Please check the route parameter for the profile name: {name}.");
+                if (profiles == null || profiles.Count < 1) return BadRequest($"Invalid request.'Profiles' property cannot be null or empty: {profiles}.");
                 var errorMessage = await _profileLogic.DeleteToolAssociations(name, profiles);
-                if (errorMessage == null)
-                {
-                    return NoContent();
-                }
-                return NotFound(errorMessage);
+                if (errorMessage == null) return NoContent();
+                else return NotFound(errorMessage);
             }
             catch (HttpRequestException ex)
             {
@@ -208,16 +167,10 @@ namespace OpenAICustomFunctionCallingAPI.Controllers
         {
             try
             {
-                if (string.IsNullOrEmpty(name))
-                {
-                    return BadRequest($"Invalid request.Please check the route parameter for the profile name: {name}.");
-                }
+                if (string.IsNullOrEmpty(name)) return BadRequest($"Invalid request.Please check the route parameter for the profile name: {name}.");
                 var success = await _profileLogic.DeleteTool(name);
-                if (success)
-                {
-                    return NoContent();
-                }
-                return NotFound($"No tool with the name {name} exists");
+                if (success) return NoContent();
+                else return NotFound($"No tool with the name {name} exists");
             }
             catch (HttpRequestException ex)
             {
